lepet/requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==4.2.16
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
frozenlist==1.4.1
phonenumbers==8.13.46
pillow==10.4.0
psycopg==3.2.2
psycopg2==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
transliterate==1.10.2
twilio==9.3.2
